control_ai/requirements.txt
numpy==1.26.4
gym==0.26.2
stable-baselines3==2.7.1
kafka-python==2.0.2
jsonschema==4.26.0
torch==2.10.0
pandas==3.0.1
pytest==8.2.2
pyyaml==6.0.1